refactor(profile): move stats/profiling hooks out of corealloc.h
Lift all heap-profile and allocator-stats logic out of the critical
corealloc.h into two new headers, behind function calls that inline to
nothing when the feature flags are off:

- mem/alloc_stats.h: AllocStats bundles FrontendStats (SNMALLOC_STATS_BASIC)
  and SizeClassStats (SNMALLOC_STATS_FULL) plus the per-Allocator hook
  methods (on_small_refill, on_local_dealloc, on_remote_ingest, ...).
  Empty and zero-size ([[no_unique_address]]) when no stats tier is built.
- profile/hooks.h: record_dealloc / record_dealloc_peek declared
  unconditionally; empty inline no-ops when SNMALLOC_PROFILE is off (the
  real definitions stay in profile/record.h).

corealloc.h drops from 21 stats/profile #ifdef regions to zero: it holds a
single AllocStats member and 11 one-line hook calls (~770 net lines
removed). stats_export.cc follows the member rename; the thread-teardown
drain call is now unconditional.

Flag-off codegen is opcode-multiset-identical to the prior baseline (zero
added or removed instructions; only register scheduling differs). Builds
clean with no flag, SNMALLOC_STATS_BASIC, SNMALLOC_STATS_FULL and
SNMALLOC_PROFILE.
